fast! versions of the Grass and Lawn interpreters
Made some new versions of the Grass and Lawn interpreters that save a ton of time and memory because they evaluate directly instead of doing lots of silly things with lambda calculus libraries. Additionally, there is a converter of Lawn programs to Grass programs (where the primitive w is substituted for primitive 0), for easy Grass coding.